requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Environment and Configuration
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1

# Database and ORM
sqlalchemy==2.0.23

# LangChain and AI Components (Stable versions)
langchain-core==0.2.39
langchain-openai==0.1.25
langchain-community==0.2.16
langgraph==0.0.26

# LlamaIndex Components (Optional - comment out if causing issues)
llama-index-core==0.10.65
llama-index-embeddings-openai==0.1.11
llama-index-llms-openai==0.1.29

# OpenAI and Token Management
openai==1.45.0
tiktoken==0.7.0

# Data Processing and Analysis
numpy==1.24.4
pandas==2.0.3
setuptools==68.2.2

# Machine Learning (Stable version)
scikit-learn==1.3.2

# Visualization
matplotlib==3.7.5

# HTTP Requests and Utilities
requests==2.31.0

# Pydantic for Data Validation
pydantic==2.4.2

# Optional Components (Comment out if not needed)
# Redis for caching
redis==4.6.0

# SQL parsing
sqlparse==0.4.4

# Async processing
celery==5.3.4

# Streamlit for UI
streamlit==1.28.2

# Development Dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1

# Additional utilities
graphviz==0.20.1

# Database drivers (install only if needed)
# psycopg2-binary==2.9.7  # For PostgreSQL
# PyMySQL==1.1.0          # For MySQL